requirements.txt
pandas
numpy
tweet-preprocessor
ftfy
clean-text[gpl]
tqdm
sklearn
scipys